services/service-patient/requirements.txt
Django==2.2.28
djangorestframework==3.12.4
django-cors-headers==3.10.1
psycopg2-binary==2.9.3
django-redis==5.2.0
redis==4.3.4
gunicorn==20.1.0
python-dotenv==0.21.1
